extract coordinate utility functions
